refactor: note의 header 부분 폴더 구조 변경
