+++ src/app/(main)/note/[id]/_components/line-info/line-info.tsx
