+++ src/app/(main)/note/[id]/_components/note-header/note-header.tsx
@@ -1,0 +1,15 @@
+import IconSelector from './icon-selector';
+import Tag from './tag';
+import Title from './title';
+
+const NoteHeader = () => {
+  return (
+    <>
+      <IconSelector />
+      <Title />
+      <Tag />
+    </>
+  );
+};
+
+export default NoteHeader;
